Fixed warning

+++ src/distribution/communication.ss/src/DeliveryHandler.h
@@ -538,10 +538,6 @@ namespace Com
 
         inline void SendAck(NodeInfo& ni, const MessageHeader* header)
         {
-//            Channel& ch=ni.GetChannel(header);
-//            auto ackPtr=boost::make_shared<Ack>(m_myId, header->commonHeader.senderId, ch.lastInSequence, header->sendMethod);
-//            WriterType::SendTo(ackPtr, ni.endpoint);
-
             Channel& ch=ni.GetChannel(header);
             auto ackPtr=boost::make_shared<Ack>(m_myId, header->commonHeader.senderId, ch.biggestSequence, header->sendMethod);
 
@@ -550,7 +546,7 @@ namespace Com
             {
                 if (seq>ch.lastInSequence)
                 {
-                    size_t index=seq-ch.lastInSequence-1;
+                    size_t index=static_cast<size_t>(seq-ch.lastInSequence-1);
                     ackPtr->missing[i]=(ch.queue[index].free ? 1 : 0);
                     --seq;
                 }
@@ -561,11 +557,6 @@ namespace Com
             }
 
             WriterType::SendTo(ackPtr, ni.endpoint);
-
-            //                static uint32_t count=0;
-            //                ++count;
-            //                if (count%1000==0)
-            //                    std::cout<<"Sent acks: "<<count<<std::endl;
         }
 
         inline boost::shared_ptr<char[]> MakePtr(const char* data, size_t size)
